add comments, move Icons folder to components
