+++ index.html
@@ -9,6 +9,10 @@
     />
     <title>Stickings Builder</title>
     <style type="text/css">
+      * {
+        box-sizing: border-box;
+        margin: 0;
+      }
       .loader {
         position: fixed;
         z-index: 2;
