+++ src/components/Icons/MetronomeIcon.tsx
