+++ src/components/Icons/Snare.tsx
