+++ src/components/Icons/SnareIcon.tsx
