+++ src/components/Icons/TripletsIcon.tsx
