+++ src/hooks/useSamples.ts
@@ -5,6 +5,9 @@ import clickLow from '../assets/audio/clickLow.wav';
 import snareR from '../assets/audio/snareR.wav';
 import snareL from '../assets/audio/snareL.wav';
 
+/* Loads sound files and creates Sampler objecs. Called on MainWrapper.tsx to ensure
+files are ready once the Metronome component is mounted*/
+
 export type Samples = {
   clickSampler: Sampler | null;
   snareSampler: Sampler | null;
@@ -33,5 +36,6 @@ export function useSamples() {
       snareSampler.dispose();
     };
   }, []);
+
   return samples;
 }
