+++ src/index.css
@@ -94,6 +94,7 @@ input[type='checkbox']:disabled {
 input[type='radio'] {
   -webkit-appearance: none;
   appearance: none;
+  margin: 0 0.2em;
   width: 20px;
   height: 20px;
   border-radius: 50%;
